requirements.txt
langchain-ollama>=0.1.0
langchain>=0.1.0
PyPDF2>=3.0.0
python-docx>=1.0.0
